// src/shared/api/types.ts
export interface ApiUserSearchResult {
  id: string;
  username: string;
  displayName: string;
  avatarUrl: string | null;
  avatarColor?: string;
}